+++ .dockerignore
@@ -1,0 +1,7 @@
+V1*
+bin*
+Docker*
+obj*
+*.json
+.*
+README*
