+++ .gitignore
@@ -1,36 +1,12 @@
-# Compiled object files
-*.o
-*.opp
-
-# Compiled static libraries
-*.a
-
-## Ignore Visual Studio temporary files, build results, and
-## files generated by popular Visual Studio add-ons.
-
-# User-specific files
-*.suo
-*.user
-*.sln.docstates
-
-# Build results
-
-build/
-[Dd]ebug/
-[Rr]elease/
-x64/
-[Bb]in/
-[Oo]bj/
-
-# Enable "build/" folder in the NuGet Packages folder since NuGet packages use it for MSBuild targets
-!packages/*/build/
-
-# MSTest test Results
-[Tt]est[Rr]esult*/
-[Bb]uild[Ll]og.*
+# .NET Core
+project.lock.json
+project.fragment.lock.json
+artifacts/
+**/Properties/launchSettings.json
 
 *_i.c
 *_p.c
+*_i.h
 *.ilk
 *.meta
 *.obj
@@ -50,168 +26,5 @@ x64/
 *.vssscc
 .builds
 *.pidb
-*.log
+*.svclog
 *.scc
-
-# Visual C++ cache files
-ipch/
-*.aps
-*.ncb
-*.opensdf
-*.sdf
-*.cachefile
-
-# Visual Studio profiler
-*.psess
-*.vsp
-*.vspx
-
-# Guidance Automation Toolkit
-*.gpState
-
-# ReSharper is a .NET coding add-in
-_ReSharper*/
-*.[Rr]e[Ss]harper
-
-# TeamCity is a build add-in
-_TeamCity*
-
-# DotCover is a Code Coverage Tool
-*.dotCover
-
-# NCrunch
-*.ncrunch*
-.*crunch*.local.xml
-
-# Installshield output folder
-[Ee]xpress/
-
-# DocProject is a documentation generator add-in
-DocProject/buildhelp/
-DocProject/Help/*.HxT
-DocProject/Help/*.HxC
-DocProject/Help/*.hhc
-DocProject/Help/*.hhk
-DocProject/Help/*.hhp
-DocProject/Help/Html2
-DocProject/Help/html
-
-# Click-Once directory
-publish/
-
-# Publish Web Output
-*.Publish.xml
-
-# NuGet Packages Directory
-packages/
-
-# Windows Azure Build Output
-csx
-*.build.csdef
-
-# Windows Store app package directory
-AppPackages/
-
-# Others
-sql/
-*.Cache
-ClientBin/
-[Ss]tyle[Cc]op.*
-~$*
-*~
-*.dbmdl
-*.[Pp]ublish.xml
-*.publishsettings
-
-# RIA/Silverlight projects
-Generated_Code/
-
-# Backup & report files from converting an old project file to a newer
-# Visual Studio version. Backup files are not needed, because we have git ;-)
-_UpgradeReport_Files/
-Backup*/
-UpgradeLog*.XML
-UpgradeLog*.htm
-
-# SQL Server files
-App_Data/*.mdf
-App_Data/*.ldf
-
-
-#LightSwitch generated files
-GeneratedArtifacts/
-_Pvt_Extensions/
-ModelManifest.xml
-
-# =========================
-# Windows detritus
-# =========================
-
-# Windows image file caches
-Thumbs.db
-ehthumbs.db
-
-# Folder config file
-Desktop.ini
-
-# Recycle Bin used on file shares
-$RECYCLE.BIN/
-
-# Mac desktop service store files
-.DS_Store
-
-# Visual studio build artifacts
-*.tlog
-*.lastbuildstate
-*.idb
-*.exp
-*.lib
-*.dll
-
-# Windows CE build artifacts
-Build.err
-Build.wrn
-Buildx86retail.dat
-*.dat
-
-# Tools EXE that doesn't end up in a typical build directory
-c/common/tools/macro_utils_h_generator/macro_utils_h_generator.exe
-
-# hg directories should be ignored
-**/hg/
-
-# Java
-**/java/**/.idea/
-**/java/**/out/
-**/java/**/target/
-**/java/**/*.iml
-**/java/**/*dependency-reduced-pom.xml
-tools/jenkins-cli.jar
-*.class
-*.jar
-**/target/
-*.iml
-
-# Node.js
-**/.settings/
-**/node_modules/
-**/.idea/
-
-# VS Code stuff
-typings/**
-.vscode/**
-
-# ignore cmake build folder
-.cmake/**
-*build/**
-build_nodejs/**
-
-# ignore Atom Editor files
-.atom-build.json
-.remote-sync.json
-
-# ignore VS Code C++ files
-browse.VC.db
-
-# api reference docstates
-api_reference/
