+++ Docker/linux-x64/Dockerfile
@@ -1,0 +1,9 @@
+FROM microsoft/dotnet:2.0.0-runtime
+
+ARG EXE_DIR=.
+
+WORKDIR /app
+
+COPY $EXE_DIR/ ./
+
+CMD ["/usr/bin/dotnet", "iot-edge-modbus.dll"]
