+++ V1/samples/modbus_sample/src/modbus_win.json
@@ -1,0 +1,99 @@
+{
+  "modules": [
+    {
+      "name": "IoTHub",
+      "loader": {
+        "name": "native",
+        "entrypoint": {
+          "module.path": "..\\..\\..\\modules\\iothub\\Debug\\iothub.dll"
+        }
+      },
+      "args": {
+        "IoTHubName": "YOUR IOT HUB NAME",
+        "IoTHubSuffix": "YOUR IOT HUB SUFFIX",
+        "Transport": "TRANSPORT PROTOCOL"
+      }
+    },
+    {
+      "name": "mapping",
+      "loader": {
+        "name": "native",
+        "entrypoint": {
+          "module.path": "..\\..\\..\\modules\\identitymap\\Debug\\identity_map.dll"
+        }
+      },
+      "args": [
+        {
+          "macAddress": "01:01:01:01:01:01",
+          "deviceId": "YOUR DEVICE ID",
+          "deviceKey": "YOUR DEVICE KEY"
+        }
+      ]
+    },
+    {
+      "name": "logger",
+      "loader": {
+        "name": "native",
+        "entrypoint": {
+          "module.path": "..\\..\\..\\modules\\logger\\Debug\\logger.dll"
+        }
+      },
+      "args": {
+        "filename": "log.txt"
+      }
+    },
+    {
+      "name": "modbus_read",
+      "loader": {
+        "name": "native",
+        "entrypoint": {
+          "module.path": "..\\..\\..\\modules\\modbus_read\\Debug\\modbus_read.dll"
+        }
+      },
+      "args": [
+        {
+          "serverConnectionString": "127.0.0.1",
+          "interval": "2000",
+          "macAddress": "01:01:01:01:01:01",
+          "baudRate": "9600",
+          "stopBits": "1",
+          "dataBits": "8",
+          "parity": "EVEN",
+          "flowControl": "NONE",
+          "deviceType": "powerMeter",
+          "sqliteEnabled": "0",
+          "operations": [
+            {
+              "unitId": "1",
+              "functionCode": "3",
+              "startingAddress": "1",
+              "length": "1"
+            }
+          ]
+        }
+      ]
+    }
+  ],
+  "links": [
+    {
+      "source": "mapping",
+      "sink": "IoTHub"
+    },
+    {
+      "source": "IoTHub",
+      "sink": "mapping"
+    },
+    {
+      "source": "mapping",
+      "sink": "modbus_read"
+    },
+    {
+      "source": "modbus_read",
+      "sink": "mapping"
+    },
+    {
+      "source": "modbus_read",
+      "sink": "logger"
+    }
+  ]
+}
