+++ iot-edge-modbus-sample.json
@@ -1,0 +1,54 @@
+﻿{
+  "SlaveConfigs": {
+    "Slave01": {
+      "SlaveConnection": "127.0.0.1",
+      "Interval": "1500",
+      "HwId": "PowerMeter-0a:01:01:01:01:01",
+      "BaudRate": "9600",
+      "StopBits": "1",
+      "DataBits": "8",
+      "Parity": "ODD",
+      "FlowControl": "NONE",
+      "SqliteEnabled": "1",
+      "Operations": {
+        "Op01": {
+          "UnitId": "1",
+          "StartAddress": "400001",
+          "Count": "2",
+          "DisplayName": "Voltage"
+        },
+        "Op02": {
+          "UnitId": "1",
+          "StartAddress": "400002",
+          "Count": "2",
+          "DisplayName": "Current"
+        }
+      }
+    },
+    "Slave02": {
+      "SlaveConnection": "127.0.0.1",
+      "Interval": "1500",
+      "HwId": "PowerMeter-0a:01:01:01:01:02",
+      "BaudRate": "9600",
+      "StopBits": "1",
+      "DataBits": "8",
+      "Parity": "ODD",
+      "FlowControl": "NONE",
+      "SqliteEnabled": "1",
+      "Operations": {
+        "Op01": {
+          "UnitId": "1",
+          "StartAddress": "40001",
+          "Count": "2",
+          "DisplayName": "Voltage"
+        },
+        "Op02": {
+          "UnitId": "1",
+          "StartAddress": "40002",
+          "Count": "2",
+          "DisplayName": "Current"
+        }
+      }
+    }
+  }
+}
