+++ iot-edge-modbus.json
@@ -1,0 +1,17 @@
+﻿{
+  "SlaveConfigs": {
+    "Slave01": {
+      "SlaveConnection": "ipv4 address to Modbus device",
+      "Interval": "read interval",
+      "HwId": "unique HW Id defined by user",
+      "Operations": {
+        "Op01": {
+          "UnitId": "unit id of Modbus device",
+          "StartAddress": "starting address of read request",
+          "Count": "unit count of read request",
+          "DisplayName": "alternative name defined by user"
+        }
+      }
+    }
+  }
+}
